+++ lib/src/list.dart
@@ -42,11 +42,17 @@ abstract class RealmList<T extends Object?> with RealmEntity implements List<T>,
   /// and it's parent object hasn't been deleted.
   bool get isValid;
 
+  /// Converts this [List] to a [RealmResults].
+  RealmResults<T> asResults();
+
   factory RealmList._(RealmListHandle handle, Realm realm, RealmObjectMetadata? metadata) => ManagedRealmList._(handle, realm, metadata);
   factory RealmList(Iterable<T> items) => UnmanagedRealmList(items);
 
   /// Creates a frozen snapshot of this `RealmList`.
   RealmList<T> freeze();
+
+  /// Allows listening for changes when the contents of this collection changes.
+  Stream<RealmListChanges<T>> get changes;
 }
 
 class ManagedRealmList<T extends Object?> with RealmEntity, ListMixin<T> implements RealmList<T> {
@@ -165,6 +171,18 @@ class ManagedRealmList<T extends Object?> with RealmEntity, ListMixin<T> impleme
     final frozenRealm = realm.freeze();
     return frozenRealm.resolveList(this)!;
   }
+
+  @override
+  RealmResults<T> asResults() => RealmResultsInternal.create<T>(realmCore.resultsFromList(this), realm, metadata);
+
+  @override
+  Stream<RealmListChanges<T>> get changes {
+    if (isFrozen) {
+      throw RealmStateError('List is frozen and cannot emit changes');
+    }
+    final controller = ListNotificationsController<T>(asManaged());
+    return controller.createStream();
+  }
 }
 
 class UnmanagedRealmList<T extends Object?> extends collection.DelegatingList<T> with RealmEntity implements RealmList<T> {
@@ -181,6 +199,12 @@ class UnmanagedRealmList<T extends Object?> extends collection.DelegatingList<T>
 
   @override
   RealmList<T> freeze() => throw RealmStateError("Unmanaged lists can't be frozen");
+
+  @override
+  RealmResults<T> asResults() => throw RealmStateError("Unmanaged lists can't be converted to results");
+
+  @override
+  Stream<RealmListChanges<T>> get changes => throw RealmStateError("Unmanaged lists don't support changes");
 }
 
 // The query operations on lists, as well as the ability to subscribe for notifications,
@@ -194,20 +218,8 @@ extension RealmListOfObject<T extends RealmObjectBase> on RealmList<T> {
   /// The Realm Dart and Realm Flutter SDKs supports querying based on a language inspired by [NSPredicate](https://academy.realm.io/posts/nspredicate-cheatsheet/)
   /// and [Predicate Programming Guide.](https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/Predicates/AdditionalChapters/Introduction.html#//apple_ref/doc/uid/TP40001789)
   RealmResults<T> query(String query, [List<Object> arguments = const []]) {
-    final managedList = asManaged();
-    final handle = realmCore.queryList(managedList, query, arguments);
+    final handle = realmCore.queryList(asManaged(), query, arguments);
     return RealmResultsInternal.create<T>(handle, realm, _metadata);
-  }
-
-  /// Allows listening for changes when the contents of this collection changes.
-  Stream<RealmListChanges<T>> get changes {
-    if (isFrozen) {
-      throw RealmStateError('List is frozen and cannot emit changes');
-    }
-
-    final managedList = asManaged();
-    final controller = ListNotificationsController<T>(managedList);
-    return controller.createStream();
   }
 }
 
@@ -275,7 +287,7 @@ extension RealmListInternal<T extends Object?> on RealmList<T> {
 }
 
 /// Describes the changes in a Realm results collection since the last time the notification callback was invoked.
-class RealmListChanges<T extends Object> extends RealmCollectionChanges {
+class RealmListChanges<T extends Object?> extends RealmCollectionChanges {
   /// The collection being monitored for changes.
   final RealmList<T> list;
 
@@ -283,7 +295,7 @@ class RealmListChanges<T extends Object> extends RealmCollectionChanges {
 }
 
 /// @nodoc
-class ListNotificationsController<T extends Object> extends NotificationsController {
+class ListNotificationsController<T extends Object?> extends NotificationsController {
   final ManagedRealmList<T> list;
   late final StreamController<RealmListChanges<T>> streamController;