+++ lib/src/results.dart
@@ -15,7 +15,6 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////////
-
 import 'dart:async';
 import 'dart:collection' as collection;
 import 'dart:ffi';
@@ -29,7 +28,7 @@ import 'realm_object.dart';
 /// added to or deleted from the Realm that match the underlying query.
 ///
 /// {@category Realm}
-class RealmResults<T extends RealmObjectBase> extends collection.IterableBase<T> with RealmEntity implements Finalizable {
+class RealmResults<T extends Object?> extends collection.IterableBase<T> with RealmEntity implements Finalizable {
   final RealmObjectMetadata? _metadata;
   final RealmResultsHandle _handle;
 
@@ -41,17 +40,13 @@ class RealmResults<T extends RealmObjectBase> extends collection.IterableBase<T>
 
   /// Returns the element of type `T` at the specified [index].
   T operator [](int index) {
-    final handle = realmCore.getObjectAt(this, index);
-    return realm.createObject(T, handle, _metadata!) as T;
-  }
-
-  /// Returns a new [RealmResults] filtered according to the provided query.
-  ///
-  /// The Realm Dart and Realm Flutter SDKs supports querying based on a language inspired by [NSPredicate](https://academy.realm.io/posts/nspredicate-cheatsheet/)
-  /// and [Predicate Programming Guide.](https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/Predicates/AdditionalChapters/Introduction.html#//apple_ref/doc/uid/TP40001789)
-  RealmResults<T> query(String query, [List<Object> args = const []]) {
-    final handle = realmCore.queryResults(this, query, args);
-    return RealmResultsInternal.create<T>(handle, realm, _metadata);
+    if (this is RealmResults<RealmObjectBase>) {
+      final handle = realmCore.resultsGetObjectAt(this, index);
+      final accessor = RealmCoreAccessor(metadata, realm.isInMigration);
+      return RealmObjectInternal.create(T, realm, handle, accessor) as T;
+    } else {
+      return realmCore.resultsGetElementAt(this, index) as T;
+    }
   }
 
   /// `true` if the `Results` collection is empty.
@@ -73,6 +68,16 @@ class RealmResults<T extends RealmObjectBase> extends collection.IterableBase<T>
   @override
   int get length => realmCore.getResultsCount(this);
 
+  /// Creates a frozen snapshot of this query.
+  RealmResults<T> freeze() {
+    if (isFrozen) {
+      return this;
+    }
+
+    final frozenRealm = realm.freeze();
+    return frozenRealm.resolveResults(this);
+  }
+
   /// Allows listening for changes when the contents of this collection changes.
   Stream<RealmResultsChanges<T>> get changes {
     if (isFrozen) {
@@ -82,15 +87,17 @@ class RealmResults<T extends RealmObjectBase> extends collection.IterableBase<T>
     final controller = ResultsNotificationsController<T>(this);
     return controller.createStream();
   }
+}
 
-  /// Creates a frozen snapshot of this query.
-  RealmResults<T> freeze() {
-    if (isFrozen) {
-      return this;
-    }
-
-    final frozenRealm = realm.freeze();
-    return frozenRealm.resolveResults(this);
+// The query operations on results only work for results of objects (core restriction),
+// so we add it as an extension methods to allow the compiler to prevent misuse.
+extension RealmResultsOfObject<T extends RealmObjectBase> on RealmResults<T> {
+  /// Returns a new [RealmResults] filtered according to the provided query.
+  ///
+  /// The Realm Dart and Realm Flutter SDKs supports querying based on a language inspired by [NSPredicate](https://www.mongodb.com/docs/realm/realm-query-language/)
+  RealmResults<T> query(String query, [List<Object> args = const []]) {
+    final handle = realmCore.queryResults(this, query, args);
+    return RealmResultsInternal.create<T>(handle, realm, _metadata);
   }
 }
 
@@ -110,15 +117,18 @@ extension RealmResultsInternal on RealmResults {
     return _handle;
   }
 
-  RealmObjectMetadata? get metadata => _metadata;
+  RealmObjectMetadata get metadata => _metadata!;
 
-  static RealmResults<T> create<T extends RealmObjectBase>(RealmResultsHandle handle, Realm realm, RealmObjectMetadata? metadata) {
-    return RealmResults<T>._(handle, realm, metadata);
-  }
+  static RealmResults<T> create<T extends Object?>(
+    RealmResultsHandle handle,
+    Realm realm,
+    RealmObjectMetadata? metadata,
+  ) =>
+      RealmResults<T>._(handle, realm, metadata);
 }
 
 /// Describes the changes in a Realm results collection since the last time the notification callback was invoked.
-class RealmResultsChanges<T extends RealmObjectBase> extends RealmCollectionChanges {
+class RealmResultsChanges<T extends Object?> extends RealmCollectionChanges {
   /// The results collection being monitored for changes.
   final RealmResults<T> results;
 
@@ -126,7 +136,7 @@ class RealmResultsChanges<T extends RealmObjectBase> extends RealmCollectionChan
 }
 
 /// @nodoc
-class ResultsNotificationsController<T extends RealmObjectBase> extends NotificationsController {
+class ResultsNotificationsController<T extends Object?> extends NotificationsController {
   final RealmResults<T> results;
   late final StreamController<RealmResultsChanges<T>> streamController;
 
@@ -158,7 +168,7 @@ class ResultsNotificationsController<T extends RealmObjectBase> extends Notifica
   }
 }
 
-class _RealmResultsIterator<T extends RealmObjectBase> implements Iterator<T> {
+class _RealmResultsIterator<T extends Object?> implements Iterator<T> {
   final RealmResults<T> _results;
   int _index;
   T? _current;